@@ -3,7 +3,7 @@ description: Demonstrates how to use the flutter_video_cast plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -12,8 +12,6 @@ dependencies:
   flutter_video_cast:
     path: ../
 
-  cupertino_icons: ^0.1.3
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -21,4 +19,3 @@ dev_dependencies:
 flutter:
 
   uses-material-design: true
-
