@@ -4,14 +4,14 @@ version: 1.0.3
 homepage: https://github.com/PalaTeam/flutter_video_cast
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  stream_transform: ^1.2.0
+  stream_transform: ^2.0.0
 
 dev_dependencies:
   flutter_test:
