@@ -20,18 +20,17 @@ abstract class AirPlayPlatform {
   }
 
   /// The route is opening.
-  Stream<RoutesOpeningEvent> onRoutesOpening({@required int id}) {
+  Stream<RoutesOpeningEvent> onRoutesOpening({required int id}) {
     throw UnimplementedError('onRoutesOpening() has not been implemented.');
   }
 
   /// The route has closed.
-  Stream<RoutesClosedEvent> onRoutesClosed({@required int id}) {
+  Stream<RoutesClosedEvent> onRoutesClosed({required int id}) {
     throw UnimplementedError('onRoutesClosed() has not been implemented.');
   }
 
   /// Returns a widget displaying the button.
-  Widget buildView(Map<String, dynamic> arguments,
-      PlatformViewCreatedCallback onPlatformViewCreated) {
+  Widget buildView(Map<String, dynamic> arguments, PlatformViewCreatedCallback onPlatformViewCreated) {
     throw UnimplementedError('buildView() has not been implemented.');
   }
 }
