@@ -7,23 +7,21 @@ part of flutter_video_cast;
 typedef void OnButtonCreated(ChromeCastController controller);
 
 /// Callback method for when a request has failed.
-typedef void OnRequestFailed(String error);
+typedef void OnRequestFailed(String? error);
 
 /// Widget that displays the ChromeCast button.
 class ChromeCastButton extends StatelessWidget {
   /// Creates a widget displaying a ChromeCast button.
-  ChromeCastButton(
-      {Key key,
-      this.size = 30.0,
-      this.color = Colors.black,
-      this.onButtonCreated,
-      this.onSessionStarted,
-      this.onSessionEnded,
-      this.onRequestCompleted,
-      this.onRequestFailed})
-      : assert(
-            defaultTargetPlatform == TargetPlatform.iOS ||
-                defaultTargetPlatform == TargetPlatform.android,
+  ChromeCastButton({
+    Key? key,
+    this.size = 30.0,
+    this.color = Colors.black,
+    this.onButtonCreated,
+    this.onSessionStarted,
+    this.onSessionEnded,
+    this.onRequestCompleted,
+    this.onRequestFailed,
+  })  : assert(defaultTargetPlatform == TargetPlatform.iOS || defaultTargetPlatform == TargetPlatform.android,
             '$defaultTargetPlatform is not supported by this plugin'),
         super(key: key);
 
@@ -37,19 +35,19 @@ class ChromeCastButton extends StatelessWidget {
   /// Callback method for when the button is ready to be used.
   ///
   /// Used to receive a [ChromeCastController] for this [ChromeCastButton].
-  final OnButtonCreated onButtonCreated;
+  final OnButtonCreated? onButtonCreated;
 
   /// Called when a cast session has started.
-  final VoidCallback onSessionStarted;
+  final VoidCallback? onSessionStarted;
 
   /// Called when a cast session has ended.
-  final VoidCallback onSessionEnded;
+  final VoidCallback? onSessionEnded;
 
   /// Called when a cast request has successfully completed.
-  final VoidCallback onRequestCompleted;
+  final VoidCallback? onRequestCompleted;
 
   /// Called when a cast request has failed.
-  final OnRequestFailed onRequestFailed;
+  final OnRequestFailed? onRequestFailed;
 
   @override
   Widget build(BuildContext context) {
@@ -57,7 +55,7 @@ class ChromeCastButton extends StatelessWidget {
       'red': color.red,
       'green': color.green,
       'blue': color.blue,
-      'alpha': color.alpha
+      'alpha': color.alpha,
     };
     return SizedBox(
       width: size,
@@ -67,29 +65,11 @@ class ChromeCastButton extends StatelessWidget {
   }
 
   Future<void> _onPlatformViewCreated(int id) async {
-    final ChromeCastController controller = await ChromeCastController.init(id);
-    if (onButtonCreated != null) {
-      onButtonCreated(controller);
-    }
-    if (onSessionStarted != null) {
-      _chromeCastPlatform
-          .onSessionStarted(id: id)
-          .listen((_) => onSessionStarted());
-    }
-    if (onSessionEnded != null) {
-      _chromeCastPlatform
-          .onSessionEnded(id: id)
-          .listen((_) => onSessionEnded());
-    }
-    if (onRequestCompleted != null) {
-      _chromeCastPlatform
-          .onRequestCompleted(id: id)
-          .listen((_) => onRequestCompleted());
-    }
-    if (onRequestFailed != null) {
-      _chromeCastPlatform
-          .onRequestFailed(id: id)
-          .listen((event) => onRequestFailed(event.error));
-    }
+    final controller = await ChromeCastController.init(id);
+    onButtonCreated?.call(controller);
+    _chromeCastPlatform.onSessionStarted(id: id).listen((_) => onSessionStarted?.call());
+    _chromeCastPlatform.onSessionEnded(id: id).listen((_) => onSessionEnded?.call());
+    _chromeCastPlatform.onRequestCompleted(id: id).listen((_) => onRequestCompleted?.call());
+    _chromeCastPlatform.onRequestFailed(id: id).listen((event) => onRequestFailed?.call(event.error));
   }
 }
