@@ -11,11 +11,11 @@ import 'package:stream_transform/stream_transform.dart';
 class MethodChannelChromeCast extends ChromeCastPlatform {
   // Keep a collection of id -> channel
   // Every method call passes the int id
-  final Map<int, MethodChannel> _channels = {};
+  final Map<int, MethodChannel?> _channels = {};
 
   /// Accesses the MethodChannel associated to the passed id.
-  MethodChannel channel(int id) {
-    return _channels[id];
+  MethodChannel? channel(int? id) {
+    return _channels[id!];
   }
 
   // The controller we need to broadcast the different events coming
@@ -26,47 +26,47 @@ class MethodChannelChromeCast extends ChromeCastPlatform {
   final _eventStreamController = StreamController<ChromeCastEvent>.broadcast();
 
   // Returns a filtered view of the events in the _controller, by id.
-  Stream<ChromeCastEvent> _events(int id) =>
+  Stream<ChromeCastEvent> _events(int? id) =>
       _eventStreamController.stream.where((event) => event.id == id);
 
   @override
   Future<void> init(int id) {
-    MethodChannel channel;
+    MethodChannel? channel;
     if (!_channels.containsKey(id)) {
       channel = MethodChannel('flutter_video_cast/chromeCast_$id');
       channel.setMethodCallHandler((call) => _handleMethodCall(call, id));
       _channels[id] = channel;
     }
-    return channel.invokeMethod<void>('chromeCast#wait');
+    return channel!.invokeMethod<void>('chromeCast#wait');
   }
 
   @override
-  Future<void> addSessionListener({int id}) {
-    return channel(id).invokeMethod<void>('chromeCast#addSessionListener');
+  Future<void> addSessionListener({int? id}) {
+    return channel(id)!.invokeMethod<void>('chromeCast#addSessionListener');
   }
 
   @override
-  Future<void> removeSessionListener({int id}) {
-    return channel(id).invokeMethod<void>('chromeCast#removeSessionListener');
+  Future<void> removeSessionListener({int? id}) {
+    return channel(id)!.invokeMethod<void>('chromeCast#removeSessionListener');
   }
 
   @override
-  Stream<SessionStartedEvent> onSessionStarted({int id}) {
+  Stream<SessionStartedEvent> onSessionStarted({int? id}) {
     return _events(id).whereType<SessionStartedEvent>();
   }
 
   @override
-  Stream<SessionEndedEvent> onSessionEnded({int id}) {
+  Stream<SessionEndedEvent> onSessionEnded({int? id}) {
     return _events(id).whereType<SessionEndedEvent>();
   }
 
   @override
-  Stream<RequestDidCompleteEvent> onRequestCompleted({int id}) {
+  Stream<RequestDidCompleteEvent> onRequestCompleted({int? id}) {
     return _events(id).whereType<RequestDidCompleteEvent>();
   }
 
   @override
-  Stream<RequestDidFailEvent> onRequestFailed({int id}) {
+  Stream<RequestDidFailEvent> onRequestFailed({int? id}) {
     return _events(id).whereType<RequestDidFailEvent>();
   }
 
@@ -79,37 +79,37 @@ class MethodChannelChromeCast extends ChromeCastPlatform {
   }
 
   @override
-  Future<void> play({@required int id}) {
-    return channel(id).invokeMethod<void>('chromeCast#play');
+  Future<void> play({required int id}) {
+    return channel(id)!.invokeMethod<void>('chromeCast#play');
   }
 
   @override
-  Future<void> pause({@required int id}) {
-    return channel(id).invokeMethod<void>('chromeCast#pause');
+  Future<void> pause({required int id}) {
+    return channel(id)!.invokeMethod<void>('chromeCast#pause');
   }
 
   @override
-  Future<void> seek(bool relative, double interval, {@required int id}) {
+  Future<void> seek(bool relative, double interval, {required int id}) {
     final Map<String, dynamic> args = {
       'relative': relative,
       'interval': interval
     };
-    return channel(id).invokeMethod<void>('chromeCast#seek', args);
+    return channel(id)!.invokeMethod<void>('chromeCast#seek', args);
   }
 
   @override
-  Future<void> stop({int id}) {
-    return channel(id).invokeMethod<void>('chromeCast#stop');
+  Future<void> stop({int? id}) async {
+    await channel(id)?.invokeMethod<void>('chromeCast#stop');
   }
 
   @override
-  Future<bool> isConnected({@required int id}) {
-    return channel(id).invokeMethod<bool>('chromeCast#isConnected');
+  Future<bool> isConnected({required int id}) async {
+    return await channel(id)?.invokeMethod<bool>('chromeCast#isConnected') ?? false;
   }
 
   @override
-  Future<bool> isPlaying({@required int id}) {
-    return channel(id).invokeMethod<bool>('chromeCast#isPlaying');
+  Future<bool> isPlaying({required int id}) async {
+    return await channel(id)?.invokeMethod<bool>('chromeCast#isPlaying') ?? false;
   }
 
   Future<dynamic> _handleMethodCall(MethodCall call, int id) async {
@@ -124,8 +124,7 @@ class MethodChannelChromeCast extends ChromeCastPlatform {
         _eventStreamController.add(RequestDidCompleteEvent(id));
         break;
       case 'chromeCast#requestDidFail':
-        _eventStreamController
-            .add(RequestDidFailEvent(id, call.arguments['error']));
+        _eventStreamController.add(RequestDidFailEvent(id, call.arguments['error']));
         break;
       default:
         throw MissingPluginException();
@@ -133,8 +132,7 @@ class MethodChannelChromeCast extends ChromeCastPlatform {
   }
 
   @override
-  Widget buildView(Map<String, dynamic> arguments,
-      PlatformViewCreatedCallback onPlatformViewCreated) {
+  Widget buildView(Map<String, dynamic> arguments, PlatformViewCreatedCallback onPlatformViewCreated) {
     if (defaultTargetPlatform == TargetPlatform.android) {
       return AndroidView(
         viewType: 'ChromeCastButton',
